Add nucleosynthesis recipes for various enchanted books
Add nucleosynthesis recipes for various enchanted books, for higher levels than are available via vanilla enchanting. This can already be achieved using Quark tomes and simply provides a technological solution to the problem. There are few uses for antimatter, and this could be something that is useful end-game for creating perfect tools.

I think I've balanced it fairly well, and it has the option to scale the antimatter requirements with the enchantment level, if that's something you want to make use of.